+++ internal/app/handlers/datasetediting/republish.go
@@ -68,7 +68,7 @@ func (h *Handler) Republish(w http.ResponseWriter, r *http.Request, ctx Context)
 		WithLevel("success").
 		WithBody(template.HTML("<p>Dataset was successfully republished.</p>"))
 
-	h.AddSessionFlash(r, w, *flash)
+	h.AddFlash(r, w, *flash)
 
 	w.Header().Set("HX-Redirect", r.URL.Query().Get("redirect-url"))
 }