+++ internal/app/handlers/publicationediting/department.go
@@ -93,7 +93,7 @@ func (h *Handler) CreateDepartment(w http.ResponseWriter, r *http.Request, ctx C
 		Note: AddDepartmentByOrg removes potential existing department
 		and then adds the new one at the end
 	*/
-	ctx.Publication.AddDepartmentByOrg(org)
+	ctx.Publication.AddOrganization(org)
 
 	// TODO handle validation errors
 
@@ -155,7 +155,7 @@ func (h *Handler) DeleteDepartment(w http.ResponseWriter, r *http.Request, ctx C
 	depID, _ := url.QueryUnescape(b.DepartmentID)
 	b.DepartmentID = depID
 
-	ctx.Publication.RemoveDepartment(b.DepartmentID)
+	ctx.Publication.RemoveOrganization(b.DepartmentID)
 
 	// TODO handle validation errors