+++ internal/app/handlers/publicationsearching/search.go
@@ -45,19 +45,19 @@ func (h *Handler) Search(w http.ResponseWriter, r *http.Request, ctx Context) {
 		ctx.SearchArgs.WithFilter("scope", "all")
 	}
 
-	searcher := h.PublicationSearchService.WithScope("status", "private", "public", "returned")
+	searcher := h.PublicationSearchIndex.WithScope("status", "private", "public", "returned")
 	args := ctx.SearchArgs.Clone()
 	var currentScope string
 
 	switch args.FilterFor("scope") {
 	case "created":
-		searcher = searcher.WithScope("creator.id", ctx.User.ID)
+		searcher = searcher.WithScope("creator_id", ctx.User.ID)
 		currentScope = "created"
 	case "contributed":
-		searcher = searcher.WithScope("author.id", ctx.User.ID)
+		searcher = searcher.WithScope("author_id", ctx.User.ID)
 		currentScope = "contributed"
 	case "all":
-		searcher = searcher.WithScope("creator.id|author.id", ctx.User.ID)
+		searcher = searcher.WithScope("creator_id|author_id", ctx.User.ID)
 		currentScope = "all"
 	default:
 		errorUnkownScope := fmt.Errorf("unknown scope: %s", args.FilterFor("scope"))
@@ -109,7 +109,7 @@ globalSearch(searcher)
 	for scoped searcher, regardless of choosen filters
 	Used to determine wether user has any records
 */
-func globalSearch(searcher backends.PublicationSearchService) (*models.PublicationHits, error) {
+func globalSearch(searcher backends.PublicationIndex) (*models.PublicationHits, error) {
 	globalArgs := models.NewSearchArgs()
 	globalArgs.Query = ""
 	globalArgs.Facets = nil
@@ -127,7 +127,7 @@ func (h *Handler) CurationSearch(w http.ResponseWriter, r *http.Request, ctx Con
 
 	ctx.SearchArgs.WithFacets(vocabularies.Map["publication_curation_facets"]...)
 
-	searcher := h.PublicationSearchService.WithScope("status", "private", "public", "returned")
+	searcher := h.PublicationSearchIndex.WithScope("status", "private", "public", "returned")
 	hits, err := searcher.Search(ctx.SearchArgs)
 	if err != nil {
 		h.Logger.Errorw("publication search: could not execute search", "errors", err, "user", ctx.User.ID)